Added character for degrees and added labels for Low/High temps
Added ANSI character for degrees (Octal 370) wherever there was a temperature mentioned. 
Example: If it was 32 F, it became 32° F. If it was 32 / 44 F (as in the forecast section), this was changed to 32 / 44 °F.
--------------
Added a Label for Low / High temperatures listed for the forecast section. The Label is meant to line up with the number below it, so if it is a 1 character temp for "Low" you will get "L". If it is a 2 character temp for "High" you will get "Hi". It will not choose "Hig" as an option because that looks awful.