Added Screenshots and License. Started change log.
Added some screenshots of the app. Regular View is real data, the Alerts section was canned test data that I manipulated. The screenshot from the Netherlands is thanks to a proxy at my.ftelnet.ca. 
Explicitly linked to the two LICENSE files. ISC was chosen for the whole project because that is what wego was using for it's icons, and I wanted to keep things simple. 
Started a simple change log/revision history.