Capture error data from JSON response
Grab error data from the JSON response and display something
for the user indicative of what went wrong. This IF statement
closes at the end of the forecast function, so basically it
wraps most of the program. Anyone with SYSOP level access
(LEVEL 90 or greater) will also get to see what the request
string looked like that was sent to wunderground.com. This might
help when debugging problems for sysops setting up the door.