Allow query to fall back to US Postal/ZIP or Airport Code
Add some new variables into ctrl/modopts.ini and weather.js that
allow the sysop to configure what the query will use as a fall back
when the caller's IP address shows up as private or non-routable.
This was requested for sysops who host their BBS remotely
(ex: a VPS) and want to not use a public IP as a fall back method.
Weather Underground API allows for the use of US Postal ZIP codes,
but this doesn't cover our international folks out there, so
the other option included is ICAO or IATA Airport Codes. The
example query on wunderground.com used ICAO but I tested with
a few IATA codes as well, and had good luck with most of them.
There are some oddball IATA codes, like MIL for Milan, Italy that
won't work correctly.