Clear abort flag before terminating
Script now clears the abort flag before terminating. 
Also added a true exit() at the end, just in case. 
This is in an attempt to correct an issue in Syncrhonet reported by Nightfox. 
Nightfox was seeing a problem if you press Q any time while syncWXremix
is running (i.e., at the pause at the end of the weather report, or while
you're reading the full alert), it quits back to the external programs menu
but Synchronet doesn't output any of the items for the external programs menu.
Digital Man fixed the problem on the Syncrhonet side, but this suggestion was
made as an attempt to help others that have not yet updated their SBBS install.