Add check for dialup users
Programatically detect a SEXPOTS/dial-up connection by 
checking if user.connection is a number (e.g. "28800") 
rather than a protocol string (e.g. "Telnet"). 
Add this as a check to the getQuerySuffix function so it
will fall back to either ZIP, Airport Code, or BBS IP 
Address (per the sysop's config in /sbbs/ctrl/modopts.ini).
Thanks to echicken for the second set of eyes on this!