Choose Degree Symbol based on Language Preference
Decide what type of Degree Symbol we should use based on the language chosen
in /ctrl/modopts.ini. I am assuming if you are using a language other than
English there will be a need for UTF-8 support to get the encoding to
look good. ANSI/CP437 character is Octal 370, UTF-8/ASCII character is
Decimal 176. I tried out the UTF-8 escaped javascript character itself,
which is \u00B0 but I did not have any luck, as it appears Sync will not
render it correctly. All testing was done with PuTTY release 0.66,
using UTF-8 as the Remote Character Set on the Translations tab, and with
a font of Courier (14-pt, Clear Type) on the Appearance tab.
If this is a pain point and causes issues, it might be best to remove the
Degree Symbol entirely from non-English language preferences.