Better detection of when language is undefined
Fix bug reported by Android8675 with a variation of the fix suggested
by echicken. The capture of the language variable being undefined did
not account for how the degreeSymbol was set, causing an error if
language variable was missing from /ctrl/modopts.ini. Now look for
typeof, and if undefined set the default language to English before
we start defining anything else.