Include language in the DEBUG output
Updated the DEBUG output to include the variable WXlang that
should be in the API call. Add a special DEBUG just to see
what langage was chosen (or not chosen) in /ctrl/modopts.ini
since wxlanguage.js will be catching things that are undefined
and obfuscating if there is a problem or something missing.