Add Backup GeoIP API nekudo.com for bad IP lookups
Wunderground has had problems lately, very specifically with GeoIP lookups. 
However, it appears that while GeoIP lookups are returning querynotfound, 
if you do a lookup by City/State or Latitude/Longitude, it gives you good results. 
So, due to a lengthy period of time with intermittent results for IP lookups with 
api.wunderground.com, I am introducing a backup GeoIP lookup. 
If we encounter querynotfound, attempt to use: 
http://geoip.nekudo.com/api/[ip-address] as our lookup, as this does 
not require one to get an API Key and is free to use, it is a good addition.
Usable results from nekudo are the Latitude and Longitude coordinates 
(it also gives City, but not State). Lat/Lon has a good success rate with 
Wunderground even when the API is having a hard time with GeoIP lookups. 
Hopefully I didn't break anything while trying to add this backup functionality.